% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmody.R
\name{modified_theme_for_ggplot2}
\alias{modified_theme_for_ggplot2}
\title{For ggplot2 theme: modify the theme of current input ggplot2
This function could add new theme for original ggplot item.
This added theme renew the text size, axis size, and lines thickness to generate new clearing plot}
\usage{
modified_theme_for_ggplot2(plot_variable)
}
\arguments{
\item{plot_variable}{a ggplot product you want to change its visualization}
}
\value{
A renew plot
}
\description{
For ggplot2 theme: modify the theme of current input ggplot2
This function could add new theme for original ggplot item.
This added theme renew the text size, axis size, and lines thickness to generate new clearing plot
}
